refactor(server): extract scene_authoring.py; drop dead prompt path
The two `_run_*_on_generator` helpers move to a new scene_authoring.py
module, returning typed `SceneEditResponseData` / `GenerateSceneResponseData`
directly instead of dicts — the receiver's `Pydantic(**dict)` splat goes
away too. The function-scoped `from image_gen import ...` is preserved
inside the new module so importing it from server.py stays light (the
diffusers/llama_cpp import graph still loads lazily on first use).

Also drops the dead prompt path: `WorldEngineManager.current_prompt`
and every `engine.set_prompt(...)` call across `engine_manager.py` and
the new `scene_authoring.py`. None of the shipped Waypoint models
condition on the prompt, so this code never had effect on generation.
The PromptNotif handler still flips the reset flag (a prompt change
forces a session reset) but no longer mutates engine state.

Net 195 lines deleted; server.py drops to ~1700 lines (down from 1735
at the start of the branch, despite all the new imports).

Refs #78.
