+++ server-components/server.py
@@ -89,6 +89,7 @@ from protocol import (
     rpc_err,
     rpc_ok,
 )
+from scene_authoring import run_generate_scene, run_scene_edit
 from video_recorder import RecordingProperties, VideoRecorder
 
 apply_resolved_token()
@@ -474,176 +475,6 @@ async def handle_check_seed_safety(
     save_safety_cache(state.safety_hash_cache)
 
     return rpc_ok(req.req_id, CheckSeedSafetyResponseData(is_safe=is_safe, hash=img_hash))
-
-
-def _run_generate_scene_on_generator(state: AppState, prompt: str, biome_version: str | None) -> dict:
-    """Generate a new scene from a text prompt and load it as the current seed.
-
-    Runs on the generator thread.  Uses the inpainting pipeline with a blank
-    canvas so the VLM can still sanitise the prompt.  Returns the generated
-    image (base64 JPEG) plus metadata (user prompt, sanitised prompt, model id)
-    so the client can persist it if Scene Authoring auto-save is enabled.
-
-    Image/image_gen imports are function-scoped to keep torch out of the
-    server module's import graph — the logged startup block in this file
-    imports torch once with progress reporting, and nothing at module level
-    should trigger a second uninstrumented load.
-    """
-    from PIL import Image
-
-    from image_gen import (
-        EDIT_MODEL_ID,
-        GENERATE_SCENE_SAFETY_MESSAGE_ID,
-        GeneratedSceneProperties,
-        SafetyRejectionError,
-        properties_to_jpeg_comment,
-    )
-
-    assert state.world_engine is not None
-    assert state.image_gen is not None
-    assert state.safety_checker is not None
-    world_engine = state.world_engine
-    image_gen = state.image_gen
-    safety_checker = state.safety_checker
-
-    t0 = time.perf_counter()
-
-    generated, sanitized_prompt = image_gen._generate_scene_sync(prompt, world_engine.seed_target_size)
-
-    # Safety check on the generated image
-    generated_np = world_engine._tensor_to_numpy(generated)
-    generated_pil = Image.fromarray(generated_np)
-    safety_result = safety_checker.check_pil_image(generated_pil)
-    if not safety_result["is_safe"]:
-        logger.warning(f"[GENERATE_SCENE] Safety checker rejected generated image: {safety_result['scores']}")
-        raise SafetyRejectionError(GENERATE_SCENE_SAFETY_MESSAGE_ID)
-
-    # Encode the generated image as JPEG so the client can save it to disk.
-    # Done before expanding to multiframe so we encode a single HxWx3 frame.
-    # Metadata is embedded via a JPEG COM-marker JSON blob — parallel to how
-    # video_recorder.py stuffs RecordingProperties into the MP4 comment atom.
-    properties = GeneratedSceneProperties(
-        biome_version=biome_version or "unknown",
-        image_model=EDIT_MODEL_ID,
-        user_prompt=prompt,
-        sanitized_prompt=sanitized_prompt,
-        generated_at=time.time(),
-    )
-    jpeg_buf = io.BytesIO()
-    generated_pil.save(
-        jpeg_buf,
-        format="JPEG",
-        quality=92,
-        comment=properties_to_jpeg_comment(properties),
-    )
-    image_b64 = base64.b64encode(jpeg_buf.getvalue()).decode("ascii")
-
-    # Expand to full temporal_compression for multiframe models
-    if world_engine.is_multiframe:
-        generated = generated.unsqueeze(0).expand(world_engine.temporal_compression, -1, -1, -1).contiguous()
-
-    # Reset engine with the generated frame as the new seed.
-    # Unlike scene_edit, this is a new world — update original_seed_frame
-    # so that "reset" returns to the generated scene, not the old seed.
-    world_engine.seed_frame = generated
-    world_engine.original_seed_frame = generated
-
-    def _reset_with_frame():
-        world_engine.engine.reset()
-        world_engine.engine.append_frame(generated)
-        if world_engine.has_prompt_conditioning:
-            world_engine.engine.set_prompt(world_engine.current_prompt)
-
-    world_engine.cuda_executor.submit(_reset_with_frame).result()
-
-    elapsed_ms = (time.perf_counter() - t0) * 1000
-    logger.info(f"[GENERATE_SCENE] Complete in {elapsed_ms:.0f}ms")
-    return {
-        "elapsed_ms": round(elapsed_ms),
-        "image_jpeg_base64": image_b64,
-        "user_prompt": prompt,
-        "sanitized_prompt": sanitized_prompt,
-        "image_model": EDIT_MODEL_ID,
-    }
-
-
-def _run_scene_edit_on_generator(state: AppState, prompt: str, cpu_frames: list) -> dict:
-    """Run inpainting on the generator thread, append via CUDA executor.
-
-    Takes the last subframe from the most recent gen_frame output,
-    inpaints it, expands to a full temporal_compression tensor, submits append_frame
-    to the CUDA executor (required for CUDA graph compatibility), and
-    returns preview data for the RPC response.
-    """
-    from PIL import Image
-
-    from image_gen import (
-        EDIT_APPEND_COUNT as SCENE_EDIT_APPEND_COUNT,
-    )
-    from image_gen import (
-        EDIT_RESET_WITH_FRAME as SCENE_EDIT_RESET,
-    )
-    from image_gen import (
-        SafetyRejectionError,
-    )
-
-    assert state.world_engine is not None
-    assert state.image_gen is not None
-    assert state.safety_checker is not None
-    world_engine = state.world_engine
-    image_gen = state.image_gen
-    safety_checker = state.safety_checker
-
-    last_frame_np = cpu_frames[-1]
-
-    # Encode original for debug preview
-    original_jpeg = world_engine._numpy_to_jpeg(last_frame_np)
-    original_b64 = base64.b64encode(original_jpeg).decode("ascii")
-
-    # Run inpainting (diffusers pipeline, not CUDA-graph dependent)
-    inpainted, edit_prompt = image_gen._inpaint_sync(last_frame_np, prompt, world_engine.seed_target_size)
-
-    # Encode inpainted for debug preview
-    inpainted_np = world_engine._tensor_to_numpy(inpainted)
-    preview_jpeg = world_engine._numpy_to_jpeg(inpainted_np)
-    preview_b64 = base64.b64encode(preview_jpeg).decode("ascii")
-
-    # Safety check on the inpainted result
-    inpainted_pil = Image.fromarray(inpainted_np)
-    safety_result = safety_checker.check_pil_image(inpainted_pil)
-    if not safety_result["is_safe"]:
-        logger.warning(f"[SCENE_EDIT] Safety checker rejected inpainted image: {safety_result['scores']}")
-        raise SafetyRejectionError()
-
-    # Expand to full temporal_compression for multiframe models
-    if world_engine.is_multiframe:
-        inpainted = inpainted.unsqueeze(0).expand(world_engine.temporal_compression, -1, -1, -1).contiguous()
-
-    # Apply the edited frame to the engine on the CUDA executor thread.
-    if SCENE_EDIT_RESET:
-        # Reset engine with the edited frame as the new seed
-        world_engine.seed_frame = inpainted
-
-        def _reset_with_frame():
-            world_engine.engine.reset()
-            world_engine.engine.append_frame(inpainted)
-            if world_engine.has_prompt_conditioning:
-                world_engine.engine.set_prompt(world_engine.current_prompt)
-
-        world_engine.cuda_executor.submit(_reset_with_frame).result()
-    else:
-        # Append repeatedly to strengthen the edit in the KV cache
-        def _append_repeated(f=inpainted):
-            for _ in range(SCENE_EDIT_APPEND_COUNT):
-                world_engine.engine.append_frame(f)
-
-        world_engine.cuda_executor.submit(_append_repeated).result()
-
-    return {
-        "original_jpeg_b64": original_b64,
-        "preview_jpeg_b64": preview_b64,
-        "edit_prompt": edit_prompt,
-    }
 
 
 # ============================================================================
@@ -1254,7 +1085,7 @@ async def websocket_endpoint(websocket: WebSocket, state: AppState = Depends(get
                                     preview = await asyncio.wrap_future(fut)
                                     # Generator thread still resolves the future with a dict;
                                     # step 5 (scene_authoring extraction) types it directly.
-                                    edit_response = rpc_ok(req.req_id, SceneEditResponseData(**preview))
+                                    edit_response = rpc_ok(req.req_id, preview)
                                 except Exception as e:
                                     error_id = getattr(e, "message_id", None)
                                     if error_id is not None:
@@ -1284,7 +1115,7 @@ async def websocket_endpoint(websocket: WebSocket, state: AppState = Depends(get
                                 generate_scene_request = {"prompt": prompt, "future": fut}
                                 try:
                                     data = await asyncio.wrap_future(fut)
-                                    gen_response = rpc_ok(req.req_id, GenerateSceneResponseData(**data))
+                                    gen_response = rpc_ok(req.req_id, data)
                                 except Exception as e:
                                     error_id = getattr(e, "message_id", None)
                                     if error_id is not None:
@@ -1429,7 +1260,7 @@ async def websocket_endpoint(websocket: WebSocket, state: AppState = Depends(get
                         req = generate_scene_request
                         generate_scene_request = None
                         try:
-                            data = _run_generate_scene_on_generator(state, req["prompt"], biome_version)
+                            data = run_generate_scene(state, req["prompt"], biome_version)
                             session.perceptual_frame_count = 0
                             req["future"].set_result(data)
                             # Send the generated seed as a single frame so the
@@ -1466,7 +1297,6 @@ async def websocket_endpoint(websocket: WebSocket, state: AppState = Depends(get
 
                     if prompt_pending is not None:
                         _flush_pending()
-                        world_engine.current_prompt = prompt_pending
                         prompt_pending = None
                         run_coro(reset_engine())
                         _action_logger_new_segment()
@@ -1496,7 +1326,7 @@ async def websocket_endpoint(websocket: WebSocket, state: AppState = Depends(get
                         scene_edit_request = None
                         _flush_pending()
                         try:
-                            preview = _run_scene_edit_on_generator(state, req["prompt"], last_generated_cpu_frames)
+                            preview = run_scene_edit(state, req["prompt"], last_generated_cpu_frames)
                             session.perceptual_frame_count = 0
                             if video_recorder is not None:
                                 video_recorder.note_edit(req["prompt"])
@@ -1512,7 +1342,7 @@ async def websocket_endpoint(websocket: WebSocket, state: AppState = Depends(get
                         generate_scene_request = None
                         _flush_pending()
                         try:
-                            data = _run_generate_scene_on_generator(state, req["prompt"], biome_version)
+                            data = run_generate_scene(state, req["prompt"], biome_version)
                             session.perceptual_frame_count = 0
                             req["future"].set_result(data)
                         except Exception as e:
